Create webscrape
This is my first webscraping program (and first big program in general) that I had created. I had actually found a thread on stack overflow for a webscraper and typed it all out and modified it to fit my needs. However, what differentiates this one, is that when the item is "out of stock" there is an "Out of Stock" label with a red background on the page, and when the item is actually in stock - that "Out of Stock" label is no longer there, and no where does it say "In Stock" in the HTML code. So that really threw me for a loop in trying to get that figured out.

I did ask for help on stack overflow, as I was 2 weeks into Python at the time to get the correct div and loop set correctly. Website URL has been partially redacted due to possible potential issues with any rules or issues pertaining to firearms and ammunition. I was pretty proud of myself for getting this fixed, and is now working flawlessly.