DataEntryIntoSQL
This has been a program I've been wanting to make for while where I would eventually turn this into an .exe program with a GUI to enter in data for my cars and the carparts that I buy, instead of putting them in manually through SQL. Not worth the time to create all this as it's easier to enter in the 5 categories instead, but nonetheless I wanted to test my skills and create this while learning Python.

As mentioned in the code, this is a work in progress until I am happy with it. The code can and will be changed, removed and added in the future.