API Integration
This is my first program working with APIs and integrating them into Python. It definitely was a challenge to get started and figure out how to use it. A simple program, but will ask a user for input and then show all the movies for that actor or actress.

This code will be improved over time.